+++ lib/image_providers/uri_image_provider.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:ui' as ui;
 
 import 'package:aves/services/common/services.dart';
+import 'package:aves_report/aves_report.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
@@ -64,14 +65,14 @@ class UriImage extends ImageProvider<UriImage> with EquatableMixin {
         },
       );
       if (bytes.isEmpty) {
-        throw StateError('$uri ($mimeType) loading failed');
+        throw UnreportedStateError('$uri ($mimeType) loading failed');
       }
       final buffer = await ui.ImmutableBuffer.fromUint8List(bytes);
       return await decode(buffer);
     } catch (error) {
       // loading may fail if the provided MIME type is incorrect (e.g. the Media Store may report a JPEG as a TIFF)
       debugPrint('$runtimeType _loadAsync failed with mimeType=$mimeType, uri=$uri, error=$error');
-      throw StateError('$mimeType decoding failed (page $pageId)');
+      throw UnreportedStateError('$mimeType decoding failed (page $pageId)');
     } finally {
       unawaited(chunkEvents.close());
     }