+++ lib/widgets/aves_app.dart
@@ -68,7 +68,9 @@ class AvesApp extends StatefulWidget {
     'ml', // Malayalam
     'my', // Burmese
     'or', // Odia
+    'sat', // Santali
     'sl', // Slovenian
+    'sv', // Swedish
     'th', // Thai
   }.map(Locale.new).toSet();
   static final List<Locale> supportedLocales = AppLocalizations.supportedLocales.where((v) => !_unsupportedLocales.contains(v)).toList();