+++ plugins/aves_model/lib/src/metadata/fields.dart
@@ -1,3 +1,11 @@
+enum MetadataSyntheticField {
+  aperture,
+  description,
+  exposureTime,
+  focalLength,
+  iso,
+}
+
 enum MetadataField {
   exifDate,
   exifDateOriginal,